针对电阻采样场景初步改进

+++ algo/algo_code_config.c
@@ -6,7 +6,7 @@ float vCTRL_TS = 1.0f / CTRL_FREQ;
 
 // PWM装载值
 int16_t vPWM_LOAD_VAL_I = PWM_LOAD_VAL;
-float vPWM_LOAD_VAL = (int16_t)PWM_LOAD_VAL;
+float vPWM_LOAD_VAL = PWM_LOAD_VAL;
 
 // PWM默认比较值
 int16_t vPWM_CMP_DEFAILT_VAL_I = PWM_CMP_DEFAILT_VAL;
